Add "look 30-days back" heuristic for small wikis without updates
Up to now small wikis that did not see updates in the last 30 days
required manual inspection and were marked "plausible" by overriding
the expected date of the csv's last line.

This manual inspection meant checking that the last editor was counted
for 30 days. As this manual step was required often recently, we now
automatize this process, and if for small wikis every non-empty column
of the last line holds non-increasing numbers for the last 30 days, we
treat the data plausible.

Change-Id: I1fcd66aa309a0a803d25b810ed8d4b09d0ca5e53