Remove '100+' and 'all' cohorts from active editors graph
The '100+' and 'all' cohorts are not required on analytics card #1155
and they'd all get shown all in the same color (dark blue), as the
their names contain the keyword 'total'. So we hide them.

As 'Active editors' implies the '5+' cohort by definition, and there
is now no longer confusion between the different cohorts, we remove
the appended cohort name.

Card: analytics 1555.1
Change-Id: Ia30a045dc29f29e84a87af15b13158fc07d86dcf