Adjust thresholds to pass at least for local checkouts again
Since geowiki's data repositories have been split into a private, and
a public part, they have been turned off. In the meantime
* 'Unkown's went up a bit (We defer solving that to later),
* Active Editors Total got restricted to showing only the 5+ cohort,
* sgwiki saw more editors, and
* some thresholds have been hit.

We adjust for the above items, to make monitoring pass on local
checkouts of the data repositories again.

The split of the data repositories has not yet been reflected in the
script, so monitoring does not yet pass without local checkouts.

Card: Analytics 1253
Change-Id: Ib876ed6956be5aec1a61e78a39a9a2c8c877bf84