Split committing and pushing limn files into separate steps
Pushing to gerrit seems to have failed at times. Since we pushed the
public data before committing the private data, failure to push the
public data aborted the script, and private data did not get committed
on the day it got generated. The private data got committed on the
first day of pushing to gerrit working again, but the private data's
history did not show changes for the other days.

Hence, we now commit on all data repositories, before trying to push
on the public repository.

Bug: 58578
Change-Id: I62382628d992c286e9d4a2aff2289f1e80deef84
Card: Analytics 1323