Switch data repository order when looping over them
As a local remote is more robust than gerrit, we now to push private
data (remote is local) before pushing the public data (remote is
gerrit). That way, if pushing to gerrit fails, the private data has
been pushed already, and the corresponding dashboards show the good,
updated data nonetheless.

Change-Id: Ibd3f25a9e82aaf6bcd73f90dcc0cdb93e53e320d